[override] Use opam pin for package override
The old method was not working in all cases for two reasons:

- when a package includes build instructions in the form of $pkg.opam
  files at the root, they must be used
- a package definition often includes more files that just `$pkg/opam`

The old behavior is preserved: when a package doesn't include a build
file in the root, `opam pin` will just update the source code, and use
the `$pkg.dev` build definition present in the repos.

Not sure if the `opam/opam-coq-archive/extra-dev` repository is needed
anymore.